Update pi-to-n.cpp
Program was previously printing to a set amount of digits including the characters "3." resulting in the program printing 2 less digits of Pi than the user would have intended. Fixed with printf.

Validation added, if the user enters a number less than or equal to 0, greater than 100 or anything other than an integer, the program will print an error and re-request the input.